Add files via upload
This project seeks to quantify the margin opportunity for prospect Airbnb hosts who wish to have a property as an income source. Airbnb has been around for more than 10 years (it was founded in 2008). For many people represents an opportunity to turn their properties into a source of income as short-term rentals. Historically, short-terms rentals were managed by real estate agencies, but through Airbnb property-owners are able to have direct contact with prospect customers who rent a property through the platform. It is difficult however to estimate how much income an owner could derive from a property given different variables that have direct impact on the likelihood of a property to be successful and have a high occupancy rate. One of the main features at the time of choosing a rental is the description that hosts include to "sell" the attractiveness of the property, highlighting the main advantages it has. There are existing solutions that aim to assess this yield opportunity such as Nestify and Houst. There are several examples of how Airbnb *reviews* can influence prospects for getting more clients in a rental. Typically this would be done by doing a sentiment analysis such as this example. Instead of customer-side reviews, I decided to focus on host-side inputs to conduct a more novel analysis, as I am sure most of my fellow students will focus the former instead of the latter.
Additionally, my aim is to bring in the text mining element to these numerical approaches, proving the impact of how a good description can boost the rental value.

There is a final challenge I wanted to take on with this project which is attempting to work with unstructured language data, because I consider that it is more realistic with professional projects.